Agents/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "farmwise-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk[eval]>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
]

[tool.setuptools]
packages = ["orchestrator"]
